chore: dry refactoring
